change using string type for both saksnummer and journalpost to hold real id value

+++ src/main/kotlin/no/nav/bidrag/hendelse/producer/kafka/RegistrertJournalpostMeldingProducer.kt
@@ -13,6 +13,6 @@ class RegistrertJournalpostMeldingProducer (@Autowired val kafkaTemplate: KafkaT
     private lateinit var TOPIC:String
 
     fun sendMelding(journalpostid:String, saksnummer:String) {
-        this.kafkaTemplate.send(TOPIC,RegistrerJournalpostDto(journalpostid.toLong(), saksnummer.toLong()));
+        this.kafkaTemplate.send(TOPIC,RegistrerJournalpostDto(journalpostid, saksnummer));
     }
 }
